requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-versatileimagefield==2.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
drf-flex-fields==0.8.6
importlib-metadata==2.0.0
Markdown==3.3
Pillow==8.0.0
PyJWT==1.7.1
python-magic==0.4.18
# python-magic-bin==0.4.14
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
typing==3.7.4.3
zipp==3.3.0
